requirements.txt
Flask~=2.2.2
google-generativeai==0.5.1
python-dotenv